Use proxy wiring for internal event validator
