+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/wiring/PlatformCoordinator.java
@@ -20,6 +20,7 @@ import com.swirlds.common.wiring.component.ComponentWiring;
 import com.swirlds.common.wiring.counters.ObjectCounter;
 import com.swirlds.platform.event.GossipEvent;
 import com.swirlds.platform.event.deduplication.EventDeduplicator;
+import com.swirlds.platform.event.validation.InternalEventValidator;
 import com.swirlds.platform.wiring.components.ApplicationTransactionPrehandlerWiring;
 import com.swirlds.platform.wiring.components.ConsensusRoundHandlerWiring;
 import com.swirlds.platform.wiring.components.EventCreationManagerWiring;
@@ -42,7 +43,7 @@ public class PlatformCoordinator {
      */
     private final ObjectCounter hashingObjectCounter;
 
-    private final InternalEventValidatorWiring internalEventValidatorWiring;
+    private final ComponentWiring<InternalEventValidator, GossipEvent> internalEventValidatorWiring;
     private final ComponentWiring<EventDeduplicator, GossipEvent> eventDeduplicatorWiring;
     private final EventSignatureValidatorWiring eventSignatureValidatorWiring;
     private final OrphanBufferWiring orphanBufferWiring;
@@ -74,7 +75,7 @@ public class PlatformCoordinator {
      */
     public PlatformCoordinator(
             @NonNull final ObjectCounter hashingObjectCounter,
-            @NonNull final InternalEventValidatorWiring internalEventValidatorWiring,
+            @NonNull final ComponentWiring<InternalEventValidator, GossipEvent> internalEventValidatorWiring,
             @NonNull final ComponentWiring<EventDeduplicator, GossipEvent> eventDeduplicatorWiring,
             @NonNull final EventSignatureValidatorWiring eventSignatureValidatorWiring,
             @NonNull final OrphanBufferWiring orphanBufferWiring,
@@ -118,7 +119,7 @@ public class PlatformCoordinator {
         // we just wait for the shared object counter to be empty, which is equivalent to flushing both components.
         hashingObjectCounter.waitUntilEmpty();
 
-        internalEventValidatorWiring.flushRunnable().run();
+        internalEventValidatorWiring.flush();
         eventDeduplicatorWiring.flush();
         eventSignatureValidatorWiring.flushRunnable().run();
         orphanBufferWiring.flushRunnable().run();
